+++ product/src/main/java/com/sparta/rooibos/product/application/dto/request/CreateProductApplicationRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.product.application.dto.req;
+package com.sparta.rooibos.product.application.dto.request;
 
 public record CreateProductApplicationRequest(String name, String clientId, String managedHubId) {
 }
